add -d keeprsp to not delete response files on success

+++ src/build.cc
@@ -29,6 +29,7 @@
 #include "deps_log.h"
 #include "disk_interface.h"
 #include "graph.h"
+#include "keep_rsp.h"
 #include "msvc_helper.h"
 #include "state.h"
 #include "subprocess.h"
@@ -779,7 +780,7 @@ void Builder::FinishCommand(CommandRunner::Result* result) {
 
   // Delete any left over response file.
   string rspfile = edge->GetBinding("rspfile");
-  if (!rspfile.empty())
+  if (!rspfile.empty() && !g_keep_rsp)
     disk_interface_->RemoveFile(rspfile);
 
   if (scan_.build_log()) {
